Allow fractional iterations*N for pFGLP.

+++ daoopt/source/ProgramOptions.h
@@ -90,7 +90,7 @@ public:
   bool useShiftedLabels; // use shifted labels induced by FGLP
   bool useNullaryShift; // use FGLP update that shifts maximums into a nullary function
   bool usePriority; // use prioritized FGLP update schedule
-  int ndfglp; // enables FGLP computation at every node dynamic MBE is used (# iters per node)
+  double ndfglp; // enables FGLP computation at every node dynamic MBE is used (# iters*N per node)
   double ndfglps; // enables FGLP computation at every node dynamic MBE is used (# seconds per node)
   double ndfglpt; // convergence tolerance for FGLP
